1.调整检索顺序 先预排序过滤，再补查详情，最后做质量检测 2.index / reindex 事务化，操作幂等性

+++ src/repositories/chunkRepository.ts
@@ -217,7 +217,7 @@ export class ChunkRepository {
                    1 - (embedding <=> $1::vector) AS similarity
             FROM ${env.chunksTable}
             WHERE embedding IS NOT NULL
-              AND status = $2
+              AND status = $2::smallint
         `;
 
         if (opts?.path) {
